+++ tests/plugins/queue/smtp_forward.js
@@ -22,7 +22,7 @@ function _setup (done) {
 }
 
 exports.loadingTLSConfig = {
-    'TLS enabled but no outbound config in tls.ini': function(test) {
+    'TLS enabled but no outbound config in tls.ini': function (test) {
         var plugin = new fixtures.plugin('queue/smtp_forward');
         test.expect(2);
 
@@ -68,7 +68,7 @@ exports.get_config = {
         test.expect(3);
         this.connection.transaction.rcpt_to.push(
             new Address('<matt@example.com>')
-            );
+        );
         var cfg = this.plugin.get_config(this.connection);
         test.equal(cfg.enable_tls, true);
         test.equal(cfg.one_message_per_rcpt, true);
@@ -79,7 +79,7 @@ exports.get_config = {
         test.expect(1);
         this.connection.transaction.rcpt_to.push(
             new Address('<matt@test.com>')
-            );
+        );
         test.deepEqual(this.plugin.get_config(this.connection), {
             host: '1.2.3.4',
             enable_tls: true,
@@ -92,7 +92,7 @@ exports.get_config = {
         test.expect(1);
         this.connection.transaction.rcpt_to.push(
             new Address('<matt@test1.com>')
-            );
+        );
         var cfg = this.plugin.get_config(this.connection);
         test.deepEqual(cfg, {
             host: '1.2.3.4',
@@ -105,13 +105,14 @@ exports.get_config = {
         this.connection.transaction.rcpt_to.push(
             new Address('<matt@test.com>'),
             new Address('<matt@test.com>')
-            );
+        );
         var cfg = this.plugin.get_config(this.connection);
         test.deepEqual(cfg.host, '1.2.3.4' );
         test.done();
     },
 };
 
+const hmail = { todo: { notes: {} } };
 exports.get_mx = {
     setUp : _setup,
     'returns no outbound route for undefined domains' : function (test) {
@@ -121,7 +122,7 @@ exports.get_mx = {
             test.deepEqual(mx, undefined);
             test.done();
         };
-        this.plugin.get_mx(cb, {}, 'undefined.com');
+        this.plugin.get_mx(cb, hmail, 'undefined.com');
     },
     'returns an outbound route for defined domains' : function (test) {
         test.expect(2);
@@ -134,6 +135,6 @@ exports.get_mx = {
             });
             test.done();
         };
-        this.plugin.get_mx(cb, {}, 'test.com');
+        this.plugin.get_mx(cb, hmail, 'test.com');
     },
 }